requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
braceexpand==0.1.7
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
clip-benchmark==1.6.1
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
datasets==3.0.0
decorator==5.1.1
dill==0.3.8
easyocr==1.7.1
einops==0.8.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.16.0
filetype==1.2.0
fire==0.6.0
Flask==3.0.3
Flask-Compress==1.15
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gevent==24.2.1
greenlet==3.1.0
grpcio==1.66.1
grpcio-tools==1.66.1
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.7
hyperframe==6.0.1
idna==3.10
imageio==2.35.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
-e git+https://github.com/apple/ml-mobileclip@341ef058802f0e4e5ab13c02f0cb32a3a94e367b#egg=mobileclip
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
paddleocr==2.8.1
paddlepaddle-gpu==2.6.2
pandas==2.2.2
pillow==10.4.0
pillow-avif-plugin==1.4.6
portalocker==2.10.1
protobuf==5.28.1
pyarrow==17.0.0
pyclipper==1.3.0.post5
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
pyparsing==3.1.4
pypdfium2==4.30.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.11.2
rapidfuzz==3.9.7
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
surya-ocr==0.5.0
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchinfo==1.8.0
torchvision==0.14.1
tqdm==4.66.5
transformers==4.44.2
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.3
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.0.4
xxhash==3.5.0
yarl==1.11.1
zope.event==5.0
zope.interface==7.0.3
zstandard==0.23.0
